deps: pin request@~2.79.0
This resolves #53 by rolling 'request' back to a version that does NOT
use har-validator@4 which pulls in a dependency on ajv@4.x which in turn
has dependencies on regenerator and nodent but does not list them in its
package.json.

In other words, rolling this back lets 'node proxy/build-client.js'
work which is required for `npm install -g strongloop`.