deps: replace minifyify with uglifyify
The source of #53 and strongloop/strongloop#337 appears to be a conflict
between minifyify and some dependencies that were introduced in request
v2.80.0 (see request/request#2585).

Since pinning the version of request hasn't been effective in the
context of 'npm install -g strongloop' we have to find an alternative
approach. By replacing minifyify with uglifyify we get the majority of
the minification benefits but without the errors.