use exorcist to extract sourcemaps
This requires upgrading browserify to a newer version because exorcist
is not able to extract the inline sorucemaps from bundles generated by
the version of browserify that was being used. In combination with the
previous commit that replaced minifyify with uglifyify, the minified
bundle is now approximately the same size as it was with minifyify and
the sourcemaps are properly externalized.